# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: UseCase Manager
  description: AI-powered use case extraction from workshop transcripts
  version: 0.1.0
paths:
...
  /auth/me:
    get:
      tags:
      - auth
      summary: Get Current User
      description: Get current authenticated user
      operationId: get_current_user_auth_me_get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  email:
                    type: string
                    title: Email
                  full_name:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Full Name
                  id:
                    type: string
                    format: uuid
                    title: Id
                  role:
                    type: string
                    enum:
                    - reader
                    - maintainer
                    - admin
                    title: UserRole
                  is_active:
                    type: boolean
                    title: Is Active
                  is_verified:
                    type: boolean
                    title: Is Verified
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - email
                - id
                - role
                - is_active
                - is_verified
                - created_at
                - updated_at
                title: UserResponse
      security:
      - OAuth2PasswordBearer: []
  /auth/users:
    get:
      tags:
      - auth
      summary: List Users
      description: List all team members (admin only)
      operationId: list_users_auth_users_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Users Auth Users Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
  /auth/users/{user_id}:
    get:
      tags:
      - auth
      summary: Get User
      description: Get a specific user
      operationId: get_user_auth_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  email:
                    type: string
                    title: Email
                  full_name:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Full Name
                  id:
                    type: string
                    format: uuid
                    title: Id
                  role:
                    type: string
                    enum:
                    - reader
                    - maintainer
                    - admin
                    title: UserRole
                  is_active:
                    type: boolean
                    title: Is Active
                  is_verified:
                    type: boolean
                    title: Is Verified
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - email
                - id
                - role
                - is_active
                - is_verified
                - created_at
                - updated_at
                title: UserResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
    delete:
      tags:
      - auth
      summary: Delete User
      description: Delete a user (admin only)
      operationId: delete_user_auth_users__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      responses:
        "204":
          description: Successful Response
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
  /auth/users/{user_id}/role:
    patch:
      tags:
      - auth
      summary: Update User Role
      description: Update user role (admin only)
      operationId: update_user_role_auth_users__user_id__role_patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                role:
                  type: string
                  enum:
                  - reader
                  - maintainer
                  - admin
                  title: UserRole
              type: object
              required:
              - role
              title: UserUpdateRole
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  email:
                    type: string
                    title: Email
                  full_name:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Full Name
                  id:
                    type: string
                    format: uuid
                    title: Id
                  role:
                    type: string
                    enum:
                    - reader
                    - maintainer
                    - admin
                    title: UserRole
                  is_active:
                    type: boolean
                    title: Is Active
                  is_verified:
                    type: boolean
                    title: Is Verified
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - email
                - id
                - role
                - is_active
                - is_verified
                - created_at
                - updated_at
                title: UserResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
...
components:
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login